Make RcuList generic over its locking strategy
Make RcuList use Lock trait instead of Futex directly, so different
locking strategies can be used. This required specifiying the lock
strategy for all current users of RcuList, which is only Qsbr where we
set the lock strategy to be the same as for the Qsbr lock. Setting it to
SpinLock might also be a sensible default but since this shouldn't be a
hot path it probably doesn't matter too much