Add long quiescent state to RCU api and Qsbr
Add a SleepingRcu trait for RcuHandles which are in a long quescent
state, and impl this for Qsbr by adding a QsbrSleeper struct. This is
useful for when a thread isn't using RCU for a while, but you don't want
to drop the RcuHandle and then have to recreate one later.